+++ core/rust-lib/src/lib.rs
@@ -298,6 +298,7 @@ pub fn run(context: tauri::Context<Wry>) {
             commands::wakelock_get,
             commands::get_finder_selection,
             commands::resize_file,
+            commands::optimize_file,
             commands::get_finder_automation_status,
             commands::open_finder_automation_settings,
             commands::force_reset_finder_automation_grant,